Fixed Stuff To Use as Examples
--- For parser.h ---
parsefile, parsefunctdefs, and parsefunction are all completed for the code that they should contain.

You're basically using what is to be expected to see if we have that (terminal tokens), and then continuing down with recursive calls (non terminal rules).

--- For the AST.h ---

Look at the class FunctionDef, this is a good example of how a completed function class would be. The NodeList, is to create "children" from the function def that contain in the example of FunctionDef, a token, then a nodelist of params (this could be null if there is non),  declaration list (again could be null), and then the body. This is how the Tree will be built.