Added getToken() & holdToken()
Need to correct for offset by holding onto a token for checking. Need to go through the functions to make sure that they aren't off by one but are getting or getting and holding a token where necessary.

Expression chain is messed up still. Needs to pass an expression down (I think), for creation.

R14 & R13 needs the holdToken()